41: convert the `src/api.py` module into an `src.api` package
